refactor: reorganizar lógica de redefinição de senha e criar componente ResetPasswordClient
